Python/tts/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
decorator==5.2.1
distro==1.9.0
et_xmlfile==2.0.0
frozenlist==1.5.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-texttospeech==2.25.1
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jiter==0.9.0
moviepy==2.1.2
multidict==6.2.0
numpy==2.2.4
openai==0.28.0
openpyxl==3.1.5
pandas==2.2.3
pillow==10.4.0
proglog==0.1.10
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
rsa==4.9
runs==1.2.2
six==1.17.0
sniffio==1.3.1
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
xmod==1.8.1
yarl==1.18.3
